requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# File handling
python-multipart>=0.0.6
aiofiles>=23.2.0

# PDF generation
weasyprint>=52.0

# Markdown processing
markdown>=3.5
Pygments>=2.17.0

# Templating
Jinja2>=3.1.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# AI Integration
google-generativeai>=0.3.0
Pillow>=10.0.0
